feat: Step2, Step3 상태관리 로직 커스텀 훅으로 분리
