--- src/pages/onboarding/OnboardingPage.tsx
@@ -1,12 +1,0 @@
-import MoodBoard from './components/steps/step3/MoodBoard';
-
-const OnboardingPage = () => {
-  return (
-    <div>
-      <h1>Onboarding Page</h1>
-      <MoodBoard />
-    </div>
-  );
-};
-
-export default OnboardingPage;
