+++ src/pages/onboarding/components/steps/step2/step2MockData.ts
