+++ src/pages/onboarding/hooks/useStep3InteriorTaste.hooks.ts
@@ -1,0 +1,62 @@
+import { useState } from 'react';
+import type {
+  CompletedInteriorTaste,
+  ImageGenerateSteps,
+} from '../types/funnel';
+
+export const useStep3InteriorTaste = (
+  context: ImageGenerateSteps['InteriorTaste'],
+  onNext: (data: CompletedInteriorTaste) => void
+) => {
+  const [selectedImages, setSelectedImages] = useState<number[]>([]);
+
+  // 이미지 선택/해제를 처리하는 함수
+  const handleImageSelect = (imageId: number) => {
+    setSelectedImages((prev) => {
+      const isSelected = prev.includes(imageId);
+
+      if (isSelected) {
+        // 이미 선택된 경우: 선택 해제
+        return prev.filter((id) => id !== imageId);
+      }
+      if (prev.length >= 5) {
+        return prev;
+      }
+      return [...prev, imageId];
+    });
+  };
+
+  const handleNext = () => {
+    // 디버깅용
+    const payload = {
+      houseType: context.houseType,
+      roomType: context.roomType,
+      roomSize: context.roomSize,
+      floorPlan: {
+        floorPlanId: context.floorPlan.floorPlanId,
+        isMirror: context.floorPlan.isMirror,
+      },
+      moodBoardIds: selectedImages,
+    };
+
+    console.log('선택된 퍼널 페이로드:', payload);
+
+    onNext({
+      houseType: context.houseType,
+      roomType: context.roomType,
+      roomSize: context.roomSize,
+      floorPlan: context.floorPlan,
+      moodBoardIds: selectedImages,
+    });
+  };
+
+  // 최소 1개 이상 선택
+  const isDataComplete = selectedImages.length > 0;
+
+  return {
+    selectedImages,
+    handleImageSelect,
+    handleNext,
+    isDataComplete,
+  };
+};
